% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub.R
\name{checkDB}
\alias{checkDB}
\title{Internal: check database}
\usage{
checkDB(sDBPATH, asTABLE, sDIGEST = NULL)
}
\arguments{
\item{sDBPATH}{文字列。データベースファイルへのパス}

\item{asTABLE}{文字列ベクトル。データベースのテーブル名。}

\item{sDIGEST}{文字列。ダイジェスト。nullの場合はチェックしない}
}
\value{
論理値。以下の条件を満たしたときにTRUE, そうでないときにFALSEを返す。
\itemize{
\item データベースファイルsDBPATHが実在する
\item そこにasTABLEとテーブル"digest"が含まれる
\item (sDIGESTが指定された場合のみ) その列の値はsDIGESTに一致する
}
}
\description{
Internal: データベースファイルのチェック
}
\keyword{internal}
