% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign.R
\name{checkSurvey}
\alias{checkSurvey}
\title{check assigned categories and slots in survey data}
\usage{
checkSurvey(lSURVEY, sVERBOSE = c("simple", "detail", "none"))
}
\arguments{
\item{lSURVEY}{an object of `surveydata` class.。
調査データ。\code{\link{makeSurvey}}で生成する。}

\item{sVERBOSE}{a string.
画面表示レベル。}
}
\value{
a data frame.
   割付結果とその検証。

   以下の列を持つ。
   \itemize{
   \item \code{nSubject}:                  対象者番号。1からの連番
   \item \code{bCat_(j)}:                  カテゴリjへの割付可能性
   \item \code{bCatSlot_(j)_(k)}:          カテゴリjのスロットkへの割付可能性
   \item \code{nCount_Cat_(j)}:            この対象者の割付終了時点でのカテゴリjへの割付人数
   \item \code{nCount_CatSlot_(j)_(k)}:    この対象者の割付終了時点でのカテゴリj, スロットkへの割付人数
   \item \code{nAssignCat_(c)}:            この対象者のc番目の割付カテゴリ(順序に意味はない)
   \item \code{nParentCat}:                この対象者の割付スロットが属するカテゴリ
   \item \code{nAssignSlot_(s)}:           この対象者のs番目の割付スロット(順序に意味はない)
   \item \code{bValid}:                    チェックを通過したか: 1:通過した, 0:通過しなかった
   \item \code{sCheckMsg}:                 チェックを通過しなかった理由。通過したときは""。
   }
}
\description{
調査データを受け取り、
カテゴリ・スロット割付が正しく行われていたことを確認する。
}
