% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias.R
\name{countBias}
\alias{countBias}
\title{count trials of simulations which are done so far}
\usage{
countBias(
  sDBPATH,
  sDBTABLE_SUBJECT = "subject",
  sDBTABLE_CAT = "weight_cat",
  sDBTABLE_SLOT = "weight_slot"
)
}
\arguments{
\item{sDBPATH}{文字列。シミュレーションの結果を保存するSQLite DBのフルパス。実在しなくてもよい。}

\item{sDBTABLE_SUBJECT}{文字列。割付シミュレーションの結果を保存するSQLite DBのテーブル名。}

\item{sDBTABLE_CAT}{文字列。再割付シミュレーションの結果(カテゴリ)を保存するSQLite DBのテーブル名。}

\item{sDBTABLE_SLOT}{文字列。再割付シミュレーションの結果(スロット)を保存するSQLite DBのテーブル名。}
}
\value{
整数。割付試行数。指定されたsDBPATHが存在しない場合は0となる。
}
\description{
これまでに実行されたシミュレーション試行(割付試行)の数を数える
}
\details{
sDBPATHが存在するのにsDBTABLE_SUBJECT, sDBTABLE_CAT, sDBTABLE_SLOTが存在しない場合はエラーになる。

   テーブルによって試行数が異なる場合はエラーになる。
}
