% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size.R
\name{countSize}
\alias{countSize}
\title{count trials of size simulations which are done so far}
\usage{
countSize(sDBPATH, sDBTABLE = "size")
}
\arguments{
\item{sDBPATH}{文字列。
シミュレーションの結果を保存するSQLite DBのフルパス。
実在しなくてもよい。}

\item{sDBTABLE}{文字列。
シミュレーションの結果を保存するSQLite DBのテーブル名。}
}
\value{
整数。
   試行数。指定されたsDBPATHが存在しない場合は0。
}
\description{
これまでに実行された標本サイズ予測シミュレーションの割付試行を数える
}
\details{
sDBPATHが存在するのにsDBTABLEが存在しない場合はエラー。
}
