% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight.R
\name{countWeight}
\alias{countWeight}
\title{count trials of simulations which are done so far}
\usage{
countWeight(
  sDBPATH,
  sDBTABLE_CAT = "weight_cat",
  sDBTABLE_SLOT = "weight_slot",
  sTYPE = c("retrial", "block")
)
}
\arguments{
\item{sDBPATH}{文字列。
シミュレーションの結果を保存するSQLite DBのフルパス。
実在しなくてもよい。}

\item{sDBTABLE_CAT}{文字列。
シミュレーションの結果(カテゴリ)を保存するSQLite DBのテーブル名。
実在しなくてもよい。}

\item{sDBTABLE_SLOT}{文字列。
シミュレーションの結果(スロット)を保存するSQLite DBのテーブル名。
実在しなくてもよい。}

\item{sTYPE}{文字列。返し値の種類。以下のいずれか:
\itemize{
\item \code{"retrial"}: 再割付試行数を返す
\item \code{"block"}: 再割付試行のブロック数を返す
}}
}
\value{
an integer.
   再割付試行数ないしブロック数。
   指定されたsDBPATH, sDBTABLE_CAT, sDBTABLE_SLOTが存在しない場合は0となる。
}
\description{
これまでに実行されたシミュレーション試行の数を数える
}
\details{
sDBPATHが存在し、sDBTABLE_CAT, sDBTABLE_SLOTのいずれかのみが存在する場合は
   エラーとなる。

   sDBPATHが存在し、sDBTABLE_CATとsDBTABLE_SLOTで試行数が異なる場合はエラーになる。
}
