% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign.R
\name{execAssign}
\alias{execAssign}
\title{Internal: Assign participants into 'categories' and 'slots'}
\usage{
execAssign(
  anSEQ_PERSON,
  mnSEQ_CAT_PERSON,
  mbPERSON_CAT_USE,
  lPERSON_SLOT_HIT,
  lSLOT_REQUEST,
  nCAT_MAX,
  sCAT_TYPE,
  sCAT_FILTER,
  sCAT_ORDER,
  sCAT_EXCLUDE,
  nSLOT_MAX,
  sSLOT_TYPE,
  sSLOT_FILTER,
  sSLOT_ORDER,
  sSLOT_EXCLUDE,
  bCHECKCOMPLETE = TRUE,
  sVERBOSE = c("simple", "detail", "none")
)
}
\arguments{
\item{anSEQ_PERSON}{an integer vector.
\code{anSEQ_PERSON[i]}は、\code{i}番目に調査に参加する対象者の
\code{mbPERSON_CAT_USE}上の行番号。欠損不可。}

\item{mnSEQ_CAT_PERSON}{an integer matrix.
\code{mnSEQ_CAT_PERSON[i,j]}は、\code{i}番目に調査に参加する
対象者の、カテゴリ\code{j}のスロット割付可能性を調べる際に参照すべき、
\code{lPERSON_SLOT_HIT[[j]]}の行番号。
欠損を許すが、
\code{mbPERSON_CAT_USE[anSEQ_PERSON[i],j] == 1}のとき、\code{mnSEQ_CAT_PERSON[i,j]}の
欠損は不可。}

\item{mbPERSON_CAT_USE}{an binary matrix.
対象者のカテゴリ割付可能性を表す。\code{mbPERSON_CAT_USE[i,j]}は以下を表す。
\itemize{
\item 1: 調査対象者iはカテゴリjについて割付可能。
\item 0: 調査対象者iはカテゴリjについて割付不能。
}
欠損を許すが、\code{mbPERSON_CAT_USE[anSEQ_PERSON, ]}における欠損は不可。}

\item{lPERSON_SLOT_HIT}{a list of binary matrices.
対象者のスロット割付可能性を表す。
要素jの行列の要素(i,k)は以下を表す。
\itemize{
\item 1: 調査対象者iはカテゴリjのスロットkについて割付可能。
\item 0: 調査対象者iはカテゴリjのスロットkについて割付不能。
}
欠損を許すが、
\code{mbPERSON_CAT_USE[i,j] == 1}のとき、
\code{lPERSON_SLOT_HIT[[j]][i, ]}における欠損は不可。}

\item{lSLOT_REQUEST}{a list of integer vectors.
各スロットに割り付ける対象者数の下限。
要素jのベクトルの要素kは, カテゴリjのスロットkに割り付ける対象者の下限を表す。}

\item{nCAT_MAX}{an integer.
ある対象者に割り付けるカテゴリ数の上限。}

\item{sCAT_TYPE}{a string.
カテゴリ割付タイプ。詳細はvignetteを参照。}

\item{sCAT_FILTER}{a string.
カテゴリ割付の際の絞り込み条件。詳細はvignetteを参照。}

\item{sCAT_ORDER}{a string.
カテゴリ割付の際の順序付け条件。詳細はvignetteを参照。}

\item{sCAT_EXCLUDE}{a.string
カテゴリ割付の際の除外条件。詳細はvignetteを参照。}

\item{nSLOT_MAX}{an integer.
ある対象者に割り付けるスロット数の上限。}

\item{sSLOT_TYPE}{a string.
スロット割付タイプ。指定は必須。詳細はvignetteを参照。}

\item{sSLOT_FILTER}{a string.
スロット割付の際の絞り込み条件。詳細はvignetteを参照。}

\item{sSLOT_ORDER}{a string.
スロット割付の際の順序付け条件。詳細はvignetteを参照。}

\item{sSLOT_EXCLUDE}{a string.
スロット割付の際の除外条件。指定は必須。詳細はvignetteを参照。}

\item{bCHECKCOMPLETE}{a logical.
\code{anSEQ_PERSON}を使い切っても
全スロットがクローズしなかったらエラーを発生させる}

\item{sVERBOSE}{a string.
画面表示レベル。}
}
\value{
an integer matrix.
   行は全スロットがクローズするまでの対象者。0行かもしれないことに注意。
   行名はmbPERSON_CAT_USEの該当行の行名。
   列は、左から順に以下の通り。ただし、
   カテゴリ番号とは\code{mbPERSON_CAT_USE}における列番号を指し、
   スロット番号とは\code{lPERSON_SLOT_HIT}の当該カテゴリにおける列番号を指す。
   \itemize{
   \item \code{SEQ}: 1             調査参加順(連番)
   \item \code{nPerson}:           対象者番号 (\code{mbPERSON_CAT_USE}上の行番号)
   \item \code{nCat_1}:            割付カテゴリ\code{1}のカテゴリ番号、ないし\code{NA}
   \item ...
   \item \code{nCat_}(nCAT_MAX):   割付カテゴリ\code{nCAT_MAX}のカテゴリ番号、ないし\code{NA}
   \item \code{nCat}:              割付スロットが属するカテゴリ番号、ないし\code{NA}
   \item \code{nSlot_1}:           割付スロット\code{1}のスロット番号、ないし\code{NA}
   \item ...
   \item \code{nSlot_}(nSLOT_MAX): 割付スロット\code{nSLOT_MAX}のスロット番号、ないし\code{NA}
   }
}
\description{
Internal: 対象者をカテゴリ・スロットに割り付ける。
}
\keyword{internal}
