% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial.R
\name{execRetrials}
\alias{execRetrials}
\title{Internal: run simulations of retrials}
\usage{
execRetrials(
  lSURVEY,
  sSAMPLING = c("with_replace", "without_replace", "fixed"),
  nBLOCKSIZE,
  nNUMBLOCK,
  bPARALLEL,
  sLOGFILE,
  sVERBOSE = c("simple", "detail", "none")
)
}
\arguments{
\item{lSURVEY}{a `surveydata`-class object.
調査データ。\code{\link{makeSurvey}}で生成する。}

\item{sSAMPLING}{a string.
抽出方法。以下のいずれか。
\itemize{
\item \code{with_replace}: 割付試行によって調査対象者を生成し、再割付試行では
調査対象者から復元無作為抽出する
\item \code{without_replace}: 割付試行によって調査対象者を生成し、再割付試行では
調査対象者から非復元無作為抽出する(調査参加順序を入れ替える)
\item \code{fixed}: 割付試行によって調査対象者を生成し、再割付試行では
それらの調査対象者を同じ調査参加順序で再割付する
}}

\item{nBLOCKSIZE}{an integer.
ブロック内の再割付試行数。}

\item{nNUMBLOCK}{an integer.
実行する再割付試行のブロック数。}

\item{bPARALLEL}{a logical.
並列処理するか。}

\item{sLOGFILE}{a string.
並列処理する場合のログファイル(フルパス)。NULLだとログを出さない。}

\item{sVERBOSE}{文字列。
画面表示レベル。}
}
\value{
データフレームのリスト。要素は次の通り:
   \itemize{
   \item \code{dfStat_Cat}: カテゴリ割付頻度
   \item \code{dfStat_Slot}: スロット割付頻度
   }

   カテゴリ割付頻度: 行は調査対象者x割付可能カテゴリを表す。
   列は以下のとおり(順不同):
   \itemize{
   \item \code{nBlock}:            ブロック番号
   \item \code{nBlockSize}:        ブロックサイズ(ブロック内の再割付試行数)
   \item \code{nSubject}:          調査対象者番号
   \item \code{nCat}:              カテゴリ番号
   \item \code{bAssign}            割付試行で割付が起きていたか
   \item \code{nNumRetrial}        その人が1回以上出現した再割付試行数
   \item \code{gSumProp}           その人が1回以上出現した再割付試行におけるカテゴリ割付率の合計
   \item \code{gSumSqProp}         その人が1回以上出現した再割付試行におけるカテゴリ割付率の二乗の合計
   }

   スロット割付頻度: 行は調査対象者x割付可能スロットを表す。
   割付試行において割付可能性が判明しているスロットに
   限定されることに注意(つまり、代替対象者のスロットは含まれない)
   列は以下のとおり(順不同):
   \itemize{
   \item \code{nBlock}:             ブロック番号
   \item \code{nBlockSize}:         ブロックサイズ(ブロック内の再割付試行数)
   \item \code{nSubject}:           調査対象者番号
   \item \code{nParentCat}:         スロットが属するカテゴリ番号
   \item \code{nSlot}:              スロット番号
   \item \code{bAssign}             割付試行で割付が起きていたか
   \item \code{nNumRetrial}         人xスロットが出現した試行数
   \item \code{gSumProp}            人xスロットが出現した試行における割付率の合計
   \item \code{gSumSqProp}          人xスロットが出現した試行における割付率の二乗の合計
   }
}
\description{
Internal: 再割付シミュレーションを行う.
simWeight(), simBias()からコールされる。
}
\keyword{internal}
