% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial.R
\name{execTrials}
\alias{execTrials}
\title{Internal: run simulation of assignment}
\usage{
execTrials(
  lPOP,
  lSETTING,
  nNUMTRIAL,
  bPARALLEL,
  sLOGFILE,
  sVERBOSE = c("simple", "detail", "none")
)
}
\arguments{
\item{lPOP}{`popdata`クラスのオブジェクト。
母集団データ。\code{\link{makePop}}で生成する。}

\item{lSETTING}{`assignsetting`クラスのオブジェクト。
割付のセッティング。\code{\link{makeSetting}} で生成する。
スロット名はlPOPと一致すること。}

\item{nNUMTRIAL}{整数。
割付シミュレーションの試行数。}

\item{bPARALLEL}{a logical.
並列処理するか。}

\item{sLOGFILE}{a string.
並列処理する場合のログファイル(フルパス)。NULLだとログを出さない。}

\item{sVERBOSE}{文字列。
画面表示レベル。}
}
\value{
データフレーム。
   行はある試行のある調査参加者を表す。
   列は以下のとおり(順不同):
   \itemize{
   \item \code{nSubject}:         対象者番号 (調査参加順の連番)
   \item \code{nPerson}:          母集団メンバー番号 (\code{lPOP$mbCAT}の行番号)
   \item \code{sRowname}:         \code{lPOP$mbCAT}の行名。\code{lPOP$mbCAT}に行名がない場合はas.character(SEQ)となる
   \item \code{nCat_1}:           割付カテゴリ\code{1}のカテゴリ番号(\code{lPOP$mbCAT}の列番号)、ないし\code{NA}
   \item ...
   \item \code{nCat_}(nMAXCAT):   割付カテゴリ\code{nMAXCAT}のカテゴリ番号(\code{lPOP$mbCAT}の列番号)、ないし\code{NA}
   \item \code{nCat}:             割付スロットが属するカテゴリ番号(\code{lPOP$mbCAT}の列番号)、ないし\code{NA}
   \item \code{nSlot_1}:          割付スロット\code{1}のスロット番号(\code{lPOP$mbSLOT[[nCat]]}の列番号)、ないし\code{NA}
   \item ...
   \item \code{nSlot_}(nMAXSLOT): 割付スロット\code{nMAXSLOT}のスロット番号(\code{lPOP$mbSLOT[[nCat]]}の列番号)、ないし\code{NA}
   }
}
\description{
Internal: 割付シミュレーションを行う.
simSize(), simBias()からコールされる。
}
\keyword{internal}
