% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias.R
\name{getBias}
\alias{getBias}
\title{get results of simulations to estimate bias}
\usage{
getBias(
  sDBPATH,
  sDBTABLE_SUBJECT = "subject",
  sDBTABLE_CAT = "weight_cat",
  sDBTABLE_SLOT = "weight_slot",
  sTYPE = c("cat", "slot")
)
}
\arguments{
\item{sDBPATH}{文字列。
シミュレーションの結果を保存したSQLite DBのフルパス}

\item{sDBTABLE_SUBJECT}{文字列。
シミュレーションの結果得られたカテゴリ割付頻度を保存したSQLite DBのテーブル名}

\item{sDBTABLE_CAT}{文字列。
シミュレーションの結果得られたカテゴリ割付頻度を保存したSQLite DBのテーブル名}

\item{sDBTABLE_SLOT}{文字列。
シミュレーションの結果得られたスロット割付頻度を保存したSQLite DBのテーブル名}

\item{sTYPE}{文字列。
取得する結果のタイプ。
\itemize{
\item \code{"cat"}: 試行xカテゴリ別のバイアス
\item \code{"slot"}: 試行xスロット別のバイアス
}}
}
\value{
データフレーム。

   \code{sTYPE == "cat"}のとき、行はある試行におけるあるカテゴリを表す。
   列は以下のとおり:
   \itemize{
   \item \code{nTrial}: 試行番号
   \item \code{nCat}: カテゴリ番号
   \item \code{nFreq_Cat}: 割り付けられた対象者の数
   \item \code{gVarWeight_Cat}: 割り付けられた対象者に与えられたウェイトの分散
   \item \code{gDeff_Cat}: 割り付けられた対象者の回答のデザイン効果
   \item \code{gESS_Cat}: 割り付けられた対象者の回答の実質標本サイズ
   }

   \code{sTYPE == "slot"}のとき、行はある試行におけるあるカテゴリを表す。
   列は以下のとおり:
   \itemize{
   \item \code{nTrial}: 試行番号
   \item \code{nCat}: カテゴリ番号
   \item \code{nSlot}: スロット番号
   \item \code{nFreq_Slot}: 割り付けられた調査対象者の数
   \item \code{gVarWeight_Slot}: 割り付けられた対象者に与えられたウェイトの分散
   \item \code{gDeff_Slot}: 割り付けられた対象者の回答のデザイン効果
   \item \code{gESS_Slot}: 割り付けられた対象者の回答の実質標本サイズ
   }
}
\description{
バイアス評価シミュレーションの結果を取得する
}
