% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub.R
\name{getCandidate}
\alias{getCandidate}
\title{Internal: get nSIZE elements from anCANDIDATE in the order of anSORTKEY.
Return all possible combinations if anSORTKEY has tied values.}
\usage{
getCandidate(anCANDIDATE, agSORTKEY, nSIZE, bDEBUG = FALSE)
}
\arguments{
\item{anCANDIDATE}{an integer vector.
候補。長さ0以上、重複なし。欠損不可。}

\item{agSORTKEY}{an numerical vector.
ソートキー。anCANDIDATEと同じ長さ。タイが許される。欠損不可。}

\item{nSIZE}{an integer.
取り出す要素数。0以上の整数。欠損不可}

\item{bDEBUG}{an logical.
デバッグモード}
}
\value{
a list.
    要素はある取り出し方による要素ベクトル。長さはnSIZE
}
\description{
Internal: anCANDIDATEからanSORTKEYが小さい順にnSIZE個の要素を取り出す。
anSORTKEYにタイがあるときはありうる取り出し方をすべて返す。
sub_checkcat(), sub_checkslot()からコールされる。
}
\keyword{internal}
