% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size.R
\name{getSize}
\alias{getSize}
\title{get results of simulations for forecasting sample size}
\usage{
getSize(sDBPATH, sDBTABLE = "size", sTYPE = c("subject", "cat", "slot"))
}
\arguments{
\item{sDBPATH}{文字列。
シミュレーションの結果を保存するSQLite DBのフルパス}

\item{sDBTABLE}{文字列。
シミュレーションの結果を保存するSQLite DBのテーブル名}

\item{sTYPE}{文字列。
取得する結果のタイプ。以下のいずれか:
\itemize{
\item \code{"subject"}: 標本サイズ(調査参加者数)
\item \code{"cat"}: カテゴリ票数
\item \code{"slot"}: スロット票数
}}
}
\value{
データフレーム。

   \code{sTYPE == "subject"}のとき、行は試行を表す。
   列は以下のとおり:
   \itemize{
   \item \code{nTrial}: 試行番号
   \item \code{nNum_scr}: 調査対象者
   \item \code{nNum_main}: 調査対象者のうち、スロット割付を受けた人数
   }

   \code{sTYPE == "cat"}のとき、行はある試行におけるあるカテゴリを表す。
   列は以下のとおり:
   \itemize{
   \item \code{nTrial}: 試行番号
   \item \code{nCat}: カテゴリ番号
   \item \code{nNum}: そのカテゴリを割り付けられた人数
   }

   \code{sTYPE == "slot"}のとき、行はある試行におけるあるカテゴリを表す。
   列は以下のとおり:
   \itemize{
   \item \code{nTrial}: 試行番号
   \item \code{nCat}:   カテゴリ番号
   \item \code{nSlot}:  スロット番号
   \item \code{nNum}:   そのスロットを割り付けられた人数
   }

   いずれも、抽出が行われなかった試行については出力されない。
   すべての試行で抽出が行われていない場合は0行のデータフレームとなる。
}
\description{
標本サイズ予測シミュレーションの結果を取得する
}
