% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size.R
\name{getSize_raw}
\alias{getSize_raw}
\title{get rawdata of simulations for forecasting sample size}
\usage{
getSize_raw(sDBPATH, sDBTABLE = "size", nTRIAL = NULL)
}
\arguments{
\item{sDBPATH}{文字列。
シミュレーションの結果を保存するSQLite DBのフルパス}

\item{sDBTABLE}{文字列。
シミュレーションの結果を保存するSQLite DBのテーブル名}

\item{nTRIAL}{整数。
試行番号。存在しないときは0行のデータを返す。
指定しないときは全試行を返す}
}
\value{
データフレーム。列は\code{\link{simSize}}を参照。
}
\description{
標本サイズ予測シミュレーションのローデータを取得する
}
