% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight.R
\name{getWeight}
\alias{getWeight}
\title{get results of simulations for computing weights}
\usage{
getWeight(
  lSURVEY,
  sDBPATH,
  sDBTABLE_CAT = "weight_cat",
  sDBTABLE_SLOT = "weight_slot",
  bCHECKDB = TRUE
)
}
\arguments{
\item{lSURVEY}{`surveydata`クラスのオブジェクト。
調査データ。\code{\link{makeSurvey}}で生成する。}

\item{sDBPATH}{文字列。
シミュレーションの結果を保存したSQLite DBのフルパス}

\item{sDBTABLE_CAT}{文字列。
シミュレーションの結果得られたカテゴリ割付頻度を保存したSQLite DBのテーブル名}

\item{sDBTABLE_SLOT}{文字列。
シミュレーションの結果得られたスロット割付頻度を保存したSQLite DBのテーブル名}

\item{bCHECKDB}{論理値。
指定されたSQLite DBをチェックするか。}
}
\value{
a list. 要素は次の通り。
  \itemize{
  \item \code{dfWeight_Cat}: a data frame. 行はある対象者のある割付カテゴリ
  \item \code{dfWeight_Slot}: a data frame. 行はある対象者のある割付スロット
  }
}
\description{
ウェイト算出シミュレーションの結果を取得する
}
