% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial.R
\name{makeRetrial}
\alias{makeRetrial}
\title{Internal: run execAssign() once}
\usage{
makeRetrial(
  lSURVEY,
  sSAMPLING = c("with_replace", "without_replace", "fixed"),
  sVERBOSE = c("simple", "detail", "none")
)
}
\arguments{
\item{lSURVEY}{a `surveydata`-class object.
survey data. See the document of `makeSurvey` for details.}

\item{sSAMPLING}{a string.
抽出方法。以下のいずれか。
\itemize{
\item \code{with_replace}: 割付試行によって調査対象者を生成し、再割付試行では
調査対象者から復元無作為抽出する
\item \code{without_replace}: 割付試行によって調査対象者を生成し、再割付試行では
調査対象者から非復元無作為抽出する(調査参加順序を入れ替える)
\item \code{fixed}: 割付試行によって調査対象者を生成し、再割付試行では
それらの調査対象者を同じ調査参加順序で再割付する
}}

\item{sVERBOSE}{a string.
画面表示レベル。}
}
\value{
a data frame.
   列は以下のとおり(順不同):
   \itemize{
   \item \code{SEQ}:              再割付における調査参加順(連番)
   \item \code{nSubject}:         対象者番号 (\code{lSURVEY$mbCAT}上の行番号)
   \item \code{nCat_1}:           割付カテゴリ\code{1}のカテゴリ番号、ないし\code{NA}
   \item ...
   \item \code{nCat_}(nMAXCAT):   割付カテゴリ\code{lSURVEY$nMAXCAT}のカテゴリ番号、ないし\code{NA}
   \item \code{nCat}:             割付スロットが属するカテゴリ番号、ないし\code{NA}
   \item \code{nSlot_1}:          割付スロット\code{1}のスロット番号、ないし\code{NA}
   \item ...
   \item \code{nSlot_}(nMAXSLOT): 割付スロット\code{lSURVEY$nMAXSLOT}のスロット番号、ないし\code{NA}
   }
}
\description{
Internal: 再割付試行を1試行実行する。
makeRetrialBlock()からコールされる。
}
\keyword{internal}
