% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial.R
\name{makeRetrialBlock}
\alias{makeRetrialBlock}
\title{Internal: run execAssign() repeatedly to compute weight}
\usage{
makeRetrialBlock(
  lSURVEY,
  nBLOCKSIZE,
  sSAMPLING = c("with_replace", "without_replace", "fixed"),
  bPARALLEL,
  sLOGFILE,
  sVERBOSE = c("simple", "detail", "none")
)
}
\arguments{
\item{lSURVEY}{a `surveydata`-class object.
survey data. See the document of `makeSurvey` for details.}

\item{nBLOCKSIZE}{整数。
試行ブロック内の試行数}

\item{sSAMPLING}{a string.
抽出方法。以下のいずれか。
\itemize{
\item \code{with_replace}: 割付試行によって調査対象者を生成し、再割付試行では
調査対象者から復元無作為抽出する
\item \code{without_replace}: 割付試行によって調査対象者を生成し、再割付試行では
調査対象者から非復元無作為抽出する(調査参加順序を入れ替える)
\item \code{fixed}: 割付試行によって調査対象者を生成し、再割付試行では
それらの調査対象者を同じ調査参加順序で再割付する
}}

\item{bPARALLEL}{a logical
並列処理するか}

\item{sLOGFILE}{a string.
並列処理する場合のログファイルの名前. NULLだとログを出さない}

\item{sVERBOSE}{文字列。画面表示レベル。}
}
\value{
a data frame. 行は再割付試行x調査参加順。
   列は以下のとおり(順不同):
   \itemize{
   \item \code{nRetrial}:         再割付試行番号
   \item \code{nSubject}:         対象者番号 (\code{lSURVEY$mbCAT}上の行番号)
   \item \code{SEQ}:              再割付試行における調査参加順(連番)
   \item \code{nCat_1}:           割付カテゴリ\code{1}のカテゴリ番号、ないし\code{NA}
   \item ...
   \item \code{nCat_}(nCAT_MAX):  割付カテゴリ\code{lSURVEY$lSETTING$nCAT_MAX}のカテゴリ番号、ないし\code{NA}
   \item \code{nCat}:             割付スロットが属するカテゴリ番号、ないし\code{NA}
   \item \code{nSlot_1}:          割付スロット\code{1}のスロット番号、ないし\code{NA}
   \item ...
   \item \code{nSlot_}(nSLOT_MAX): 割付スロット\code{lSURVEY$lSETTING$nSLOT_MAX}のスロット番号、ないし\code{NA}
   }
}
\description{
Internal: 再割付試行を1ブロック実行する。execRetrials()からコールされる。
}
\keyword{internal}
