% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{makeSetting}
\alias{makeSetting}
\title{make setting}
\usage{
makeSetting(
  lSLOT_REQUEST,
  nCAT_MAX,
  sCAT_ASSIGN,
  nSLOT_MAX,
  sSLOT_ASSIGN,
  nSUBJECT_MAX = 0,
  nCATSUBJECT_MAX = 0,
  nSLOTSUBJECT_MAX = 0,
  bSTOP_WHEN_FULFILLED = TRUE,
  sVERBOSE = c("simple", "none", "detail")
)
}
\arguments{
\item{lSLOT_REQUEST}{a list of (named) integer vectors.
各スロットに割り付ける対象者数の目標。
要素jのベクトルの要素kは, カテゴリjのスロットkに割り付ける対象者の下限を表す。
欠損不可。
要素に名前がついている場合、スロット名とみなされるので、
名前を付けるならばすべての要素に重複なく名前をつけること。}

\item{nCAT_MAX}{a integer.
ある対象者に割り付けるカテゴリ数の上限。}

\item{sCAT_ASSIGN}{a string.
カテゴリ割付方法。詳細はvignetteを参照。}

\item{nSLOT_MAX}{an integer.
ある対象者に割り付けるスロット数の上限。}

\item{sSLOT_ASSIGN}{a string.
スロット割付方法。詳細はvignetteを参照。}

\item{nSUBJECT_MAX}{an integer.
抽出する対象者数の上限。0以上の整数。
0より大きい値が指定された場合、対象者抽出は、
対象者数がその値に達したときに中止される。}

\item{nCATSUBJECT_MAX}{an integer.
カテゴリ割付対象者数の上限。0以上の整数。
0より大きい値が指定された場合、対象者抽出は、
カテゴリ割付が生じた対象者数がその値に達したときに中止される。}

\item{nSLOTSUBJECT_MAX}{an integer.
スロット割付対象者数の上限。0以上の整数。
0より大きい値が指定された場合、対象者抽出は、
スロット割付が生じた対象者数がその値に達したときに中止される。}

\item{bSTOP_WHEN_FULFILLED}{as boolean.
調査停止ルール。nSUBJECT_MAX, nCATSUBJECT_MAX, nSLOTSUBJECT_MAXのいずれかが
0以上である場合、これに到達した時に調査停止となる。さらに、
この引数がTRUEであるときには、
すべてのスロットにおいて割り付けられた対象者数が
目標に到達した時にも調査停止となる。}

\item{sVERBOSE}{a string.
画面表示レベル。}
}
\value{
an object of `assignsetting` class.
        その実体は、この関数の引数を要素として持つリスト。
        ただし、引数lSLOT_REQUESTの要素に名前がついていない場合は、
        名前"Slot_(j)_(k)"が付与される。
}
\description{
割付のセッティングを作成する。
}
\examples{
lSetting <- makeSetting(
  lSLOT_REQUEST = lapply(1:10, function(x) rep(100, 10)),
  nCAT_MAX      = 1,
  sCAT_ASSIGN   = 'all-random-assignable-none',
  nSLOT_MAX     = 2,
  sSLOT_ASSIGN  = 'all-random-assignable-none',
  sVERBOSE      = "detail"
)
}
