% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial.R
\name{makeStat_Slot}
\alias{makeStat_Slot}
\title{Interal: make statistics for slots}
\usage{
makeStat_Slot(
  dfReplicateBlock,
  lSURVEY,
  sVERBOSE = c("simple", "detail", "none")
)
}
\arguments{
\item{dfReplicateBlock}{a data frame
対象者のカテゴリ・ウェイトへの割付を反復実行した結果。
以下の列を持つと期待される
 \itemize{
 \item \code{nRetrial}:         再割付試行番号
 \item \code{nSubject}:         対象者番号
 \item \code{nCat_1}:           割付カテゴリ\code{1}のカテゴリ番号、ないし\code{NA}
 \item ...
 \item \code{nCat_}(nMAXCAT):   割付カテゴリ\code{nMAXCAT}のカテゴリ番号、ないし\code{NA}
 \item \code{nParentCat}:       割付スロットが属するカテゴリ番号、ないし\code{NA}
 \item \code{nSlot_1}:          割付スロット\code{1}のスロット番号、ないし\code{NA}
 \item ...
 \item \code{nSlot_}(nMAXSLOT): 割付スロット\code{nMAXSLOT}のスロット番号、ないし\code{NA}
 }}

\item{lSURVEY}{a `surveydata`-class object.
survey data. See the document of `makeSurvey` for details.}

\item{sVERBOSE}{文字列。
画面表示レベル。}
}
\value{
a data frame
   行は対象者x割付可能スロット。lSURVEY上で割付可能性が判明しているスロットに
   限定されることに注意(つまり、代替対象者のスロットは含まれない)
   列は以下の通り
   \itemize{
   \item \code{nSubject}               対象者番号
   \item \code{nParentCat}             親カテゴリ番号
   \item \code{nSlot}                  スロット番号
   \item \code{bAssign}                lSURVEY上で割付が起きていたか
   \item \code{nNumRetrial}            人xスロットが出現した試行数
   \item \code{gSumProp}               人xスロットが出現した試行における割付率の合計
   \item \code{gSumSqProp}             人xスロットが出現した試行における割付率の二乗の合計
   }
}
\description{
Internal: スロット割付頻度を集計する。execRetrials()からコールされる。
}
\keyword{internal}
