% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSurveyAssign-package.R
\docType{package}
\name{rSurveyAssign-package}
\alias{rSurveyAssign-package}
\alias{rSurveyAssign}
\title{rSurveyAssign: run simulation of assignment in a web survey}
\description{
run simulation of assignment in a web survey
}
\details{
This package contains several tools to simulate different types of assignment
processes in web surveys. With this package, users can estimate sample sizes
and for future surveys, assess bias in future surveys, and calculate
survey weights for actual surveys that have been conducted.

このパッケージは、web調査における対象者割付についてのシミュレーションを
行うためのパッケージです。これから行う調査について標本サイズを見積もったり、
バイアスを評価したり、すでに行った調査について調査ウェイトを求めたりすることができます。
}
\seealso{
Useful links:
\itemize{
  \item \url{https://shigono.github.io/rSurveyAssign/}
}

}
