% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias.R
\name{simBias}
\alias{simBias}
\title{run simulations to evaluate bias}
\usage{
simBias(
  lPOP,
  lSETTING,
  nNUMTRIAL = 1,
  nNUMRETRIAL = 1000,
  bUSE_INFO_UNASSIGNED_CAT = TRUE,
  bPARALLEL = FALSE,
  sLOGFILE = NULL,
  sDBPATH = NULL,
  sDBTABLE_SUBJECT = "subject",
  sDBTABLE_CAT = "weight_cat",
  sDBTABLE_SLOT = "weight_slot",
  bAPPEND = TRUE,
  sVERBOSE = c("simple", "detail", "none")
)
}
\arguments{
\item{lPOP}{`popdata`クラスのオブジェクト。
母集団データ。\code{\link{makePop}}で生成する。}

\item{lSETTING}{`assignsetting'クラスのオブジェクト。
割付のセッティング。\code{\link{makeSetting}}で生成する。}

\item{nNUMTRIAL}{整数。
割付試行数。}

\item{nNUMRETRIAL}{整数。
各割付試行における再割付試行数。}

\item{bUSE_INFO_UNASSIGNED_CAT}{論理値。
再割付試行において、非割付カテゴリに属するスロットへの割付可能性を利用するか。
TRUEの場合、すべての割付可能カテゴリに属するスロット割付可能性を用いて
再割付試行を行う。
FALSEの場合、割付カテゴリに属するスロット割付可能性のみを用いて
再割付試行を行う。}

\item{bPARALLEL}{論理値。
並列処理するか。}

\item{sLOGFILE}{文字列。
並列処理する場合のログファイル(フルパス)。NULLだとログを出さない。}

\item{sDBPATH}{文字列。
シミュレーションの結果を保存するSQLite DBのフルパス。}

\item{sDBTABLE_SUBJECT}{文字列。
割付シミュレーションの結果を保存するSQLite DBのテーブル名。}

\item{sDBTABLE_CAT}{文字列。
再割付シミュレーションの結果得られたカテゴリ割付頻度を保存するSQLite DBのテーブル名。}

\item{sDBTABLE_SLOT}{文字列。
再割付シミュレーションの結果得られたスロット割付頻度を保存するSQLite DBのテーブル名。}

\item{bAPPEND}{論理値。
TRUE: 結果をテーブルに追加する。FALSE: 上書きする。}

\item{sVERBOSE}{文字列。
画面表示レベル。}
}
\value{
\itemize{
     \item sDBPATHを指定しない場合は、シミュレーションの結果をデータフレームのリストとして返す。
     \item sDBPATHを指定した場合はNULLを返す。シミュレーションの結果は、
           SQLiteデータベースにテーブルとして保存される。sDBPATH が存在する場合は、
           bAPPEND==FALSEであればSDBPATHに上書きし、bAPPEND==TRUEであれば追加する。
           sDBPATHが存在しない場合は新規作成する。
   }

   シミュレーションの結果は次の3つのデータフレームからなる。
   \itemize{
   \item \code{dfSubject}: 割付シミュレーションの結果
   \item \code{dfStat_Cat}: 再割付シミュレーションにおけるカテゴリ割付頻度
   \item \code{dfStat_Slot}: 再割付シミュレーションにおけるスロット割付頻度
   }

   割付シミュレーションの結果:
   行はある割付試行のある調査参加者を表す。
   列は以下のとおり(順不同):
   \itemize{
   \item \code{nTrial}:           試行番号
   \item \code{nSubject}:         対象者番号 (調査参加順の連番)
   \item \code{nPerson}:          母集団メンバー番号 (\code{lPOP$mbCAT}の行番号)
   \item \code{sRowname}:         \code{lPOP$mbCAT}の行名。\code{lPOP$mbCAT}に行名がない場合はas.character(SEQ)となる
   \item \code{nCat_1}:           割付カテゴリ\code{1}のカテゴリ番号(\code{lPOP$mbCAT}の列番号)、ないし\code{NA}
   \item ...
   \item \code{nCat_}(nMAXCAT):   割付カテゴリ\code{nMAXCAT}のカテゴリ番号(\code{lPOP$mbCAT}の列番号)、ないし\code{NA}
   \item \code{nCat}:             割付スロットが属するカテゴリ番号(\code{lPOP$mbCAT}の列番号)、ないし\code{NA}
   \item \code{nSlot_1}:          割付スロット\code{1}のスロット番号(\code{lPOP$mbSLOT[[nCat]]}の列番号)、ないし\code{NA}
   \item ...
   \item \code{nSlot_}(nMAXSLOT): 割付スロット\code{nMAXSLOT}のスロット番号(\code{lPOP$mbSLOT[[nCat]]}の列番号)、ないし\code{NA}
   }

   再割付シミュレーションにおけるカテゴリ割付頻度:
   行は調査対象者x割付可能カテゴリを表す。
   列は以下のとおり(順不同):
   \itemize{
   \item \code{nTrial}:            試行番号
   \item \code{nBlock}:            ブロック番号
   \item \code{nBlockSize}:        ブロックサイズ(ブロック内の再割付試行数)
   \item \code{nSubject}:          調査対象者番号
   \item \code{nCat}:              カテゴリ番号
   \item \code{sRowname}:           \code{lPOP$mbCAT}の行名。\code{lPOP$mbCAT}に行名がない場合はas.character(SEQ)となる
   \item \code{nPerson}:            母集団メンバー番号 (\code{lPOP$mbCAT}の行番号)
   \item \code{bAssign}            割付試行で割付が起きていたか
   \item \code{nCount_Subject}:    ブロック内の再割付試行で調査対象者が出現した回数
   \item \code{nCount_SubjectCat}: ブロック内の再割付試行で調査対象者が出現しカテゴリに割り付けられた回数
   \item \code{gAssignablity_Cat}: カテゴリ割付可能度(全カテゴリに占める割付可能カテゴリの割合)
   }

   再割付シミュレーションにおけるスロット割付頻度:
   行は調査対象者x割付可能スロットを表す。
   割付試行において割付可能性が判明しているスロットに
   限定されることに注意(つまり、代替対象者のスロットは含まれない)
   列は以下のとおり(順不同):
   \itemize{
   \item \code{nTrial}:             試行番号
   \item \code{nBlock}:             ブロック番号
   \item \code{nBlockSize}:         ブロックサイズ(ブロック内の再割付試行数)
   \item \code{nSubject}:           調査対象者番号
   \item \code{sRowname}:           \code{lPOP$mbCAT}の行名。\code{lPOP$mbCAT}に行名がない場合はas.character(SEQ)となる
   \item \code{nPerson}:            母集団メンバー番号 (\code{lPOP$mbCAT}の行番号)
   \item \code{nCat}:               スロットが属するカテゴリ番号
   \item \code{nSlot}:              スロット番号
   \item \code{bAssign}             割付試行で割付が起きていたか
   \item \code{nCount_Subject}:     ブロック内の再割付試行で調査対象者が出現した回数
   \item \code{nCount_SubjectSlot}: ブロック内の再割付試行で調査対象者が出現しスロットに割り付けられた回数
   \item \code{gAssignablity_Slot}: スロット割付可能度(全スロットに占める割付可能スロットの割合)
   }
}
\description{
バイアス評価のためのシミュレーションを実行する
}
\details{
bAPPEND==TRUEで、sDBPATHが存在し、かつ
   テーブルsDBTABLE_SUBJECT, sDBTABLE_CAT, sDBTABLE_SLOTが存在しない場合はエラーとなる。
}
