% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size.R
\name{simSize}
\alias{simSize}
\title{run simulations for forecasting sample size}
\usage{
simSize(
  lPOP,
  lSETTING,
  nNUMTRIAL = 1,
  bPARALLEL = FALSE,
  sLOGFILE = NULL,
  sDBPATH = NULL,
  sDBTABLE = "size",
  bAPPEND = TRUE,
  sVERBOSE = c("simple", "detail", "none")
)
}
\arguments{
\item{lPOP}{`popdata`クラスのオブジェクト。
母集団データ。\code{\link{makePop}}で生成する。}

\item{lSETTING}{`assignsetting`クラスのオブジェクト。
割付のセッティング。\code{\link{makeSetting}} で生成する。}

\item{nNUMTRIAL}{整数。
割付シミュレーションの試行数。}

\item{bPARALLEL}{論理値。
並列処理するか。}

\item{sLOGFILE}{文字列。
並列処理する場合のログファイル(フルパス)。NULLだとログを出さない。}

\item{sDBPATH}{文字列。
シミュレーションの結果を保存するSQLite DBのフルパス。}

\item{sDBTABLE}{文字列。
シミュレーションの結果を保存するSQLite DBのテーブル名。}

\item{bAPPEND}{論理値。
TRUE: 結果をテーブルに追加する。FALSE: 上書きする。}

\item{sVERBOSE}{文字列。
画面表示レベル。}
}
\value{
sDBPATHを指定しなかった場合は、シミュレーションの結果を
   データフレームとして返す。

   sDBPATHを指定した場合はNULLを返す。
   シミュレーションの結果は、SQLiteデータベースにテーブルとして
   保存される。

   bAPPEND==FALSEの場合、SQLiteデータベースファイル sDBPATH
   が存在する場合は上書きし、存在しない場合は新規作成する。

   bAPPEND==TRUEの場合、SQLiteデータベースファイル sDBPATH
   が存在する場合にはテーブルsDBTABLEに追加する。
   ファイルが存在しない場合はファイルとテーブルを新規作成する。

   シミュレーションの結果は以下の形式を持つ。
   行はある試行のある調査参加者を表す。0行かもしれないことに注意。
   列は以下のとおり(順不同):
   \itemize{
   \item \code{nSubject}:         対象者番号 (調査参加順の連番)
   \item \code{nPerson}:          母集団メンバー番号 (\code{lPOP$mbCAT}の行番号)
   \item \code{sRowname}:         \code{lPOP$mbCAT}の行名。\code{lPOP$mbCAT}に行名がない場合はas.character(SEQ)となる
   \item \code{nCat_1}:           割付カテゴリ\code{1}のカテゴリ番号(\code{lPOP$mbCAT}の列番号)、ないし\code{NA}
   \item ...
   \item \code{nCat_}(nMAXCAT):   割付カテゴリ\code{nMAXCAT}のカテゴリ番号(\code{lPOP$mbCAT}の列番号)、ないし\code{NA}
   \item \code{nCat}:             割付スロットが属するカテゴリ番号(\code{lPOP$mbCAT}の列番号)、ないし\code{NA}
   \item \code{nSlot_1}:          割付スロット\code{1}のスロット番号(\code{lPOP$mbSLOT[[nCat]]}の列番号)、ないし\code{NA}
   \item ...
   \item \code{nSlot_}(nMAXSLOT): 割付スロット\code{nMAXSLOT}のスロット番号(\code{lPOP$mbSLOT[[nCat]]}の列番号)、ないし\code{NA}
   }
}
\description{
標本サイズ推定のためのシミュレーションを行う
}
\details{
bAPPEND==TRUEで、sDBPATHが存在するがテーブルsDBTABLEが存在しない場合は
エラーとなる。
}
