% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight.R
\name{simWeight}
\alias{simWeight}
\title{run simulations for computing weights}
\usage{
simWeight(
  lSURVEY,
  sSAMPLING = c("with_replace", "without_replace", "fixed"),
  nBLOCKSIZE = 100,
  nNUMBLOCK = 1,
  bPARALLEL = FALSE,
  sLOGFILE = NULL,
  sDBPATH = NULL,
  sDBTABLE_CAT = "weight_cat",
  sDBTABLE_SLOT = "weight_slot",
  bAPPEND = FALSE,
  sVERBOSE = c("simple", "detail", "none")
)
}
\arguments{
\item{lSURVEY}{`surveydata`クラスのオブジェクト。
調査データ。\code{\link{makeSurvey}}で生成する。}

\item{sSAMPLING}{a string.
抽出方法。以下のいずれか。
\itemize{
\item \code{with_replace}: 割付試行によって調査対象者を生成し、再割付試行では
調査対象者から復元無作為抽出する
\item \code{without_replace}: 割付試行によって調査対象者を生成し、再割付試行では
調査対象者から非復元無作為抽出する(調査参加順序を入れ替える)
\item \code{fixed}: 割付試行によって調査対象者を生成し、再割付試行では
それらの調査対象者を同じ調査参加順序で再割付する
}}

\item{nBLOCKSIZE}{整数。
ブロック内の再割付試行数。}

\item{nNUMBLOCK}{整数。
実行する再割付試行のブロック数。}

\item{bPARALLEL}{論理値。
並列処理するか。}

\item{sLOGFILE}{文字列。
並列処理する場合のログファイル(フルパス)。NULLだとログを出さない。}

\item{sDBPATH}{文字列。
シミュレーションの結果を保存するSQLite DBのフルパス。
存在しない場合は作成する。}

\item{sDBTABLE_CAT}{文字列。
再割付シミュレーションの結果得られたカテゴリ割付頻度を保存するSQLite DBのテーブル名。}

\item{sDBTABLE_SLOT}{文字列。
再割付シミュレーションの結果得られたスロット割付頻度を保存するSQLite DBのテーブル名。}

\item{bAPPEND}{論理値。
結果をテーブルに TRUE: 追加する, FALSE: 上書きする。}

\item{sVERBOSE}{文字列。
画面表示レベル。}
}
\value{
\itemize{
     \item sDBPATHを指定しない場合は、シミュレーションの結果をデータフレームのリストとして返す。
     \item sDBPATHを指定した場合はNULLを返す。シミュレーションの結果は、
           SQLiteデータベースにテーブルとして保存される。sDBPATH が存在する場合は、
           bAPPEND==FALSEであればSDBPATHに上書きし、bAPPEND==TRUEであれば追加する。
           sDBPATHが存在しない場合は新規作成する。
   }

   シミュレーションの結果は次の2つのデータフレームからなる。

   \code{dfStat_Cat}: カテゴリ割付頻度。行は調査対象者x割付カテゴリを表す。
   列は以下のとおり(順不同):
   \itemize{
   \item \code{nBlock}:           ブロック番号
   \item \code{nBlockSize}:       ブロックサイズ(ブロック内の再割付試行数)
   \item \code{nSubject}:         調査対象者番号
   \item \code{nCat}:             カテゴリ番号
   \item \code{nNumRetrial}:      ブロック内の再割付試行のうち、調査対象者が1回以上出現した試行の数
   \item \code{gSumProp}: 　      ブロック内の再割付試行のうち、調査対象者が1回以上出現した試行を通じた、割付率の合計
   \item \code{gSumSqProp}:       ブロック内の再割付試行のうち、調査対象者が1回以上出現した試行を通じた、割付率の二乗の合計
   }

   \code{dfStat_Slot}: スロット割付頻度。行は調査対象者x割付スロットを表す。
   列は以下のとおり(順不同):
   \itemize{
   \item \code{nBlock}:           ブロック番号
   \item \code{nBlockSize}:       ブロックサイズ(ブロック内の再割付試行数)
   \item \code{nSubject}:         調査対象者番号
   \item \code{nParentCat}:       スロットが属するカテゴリ番号
   \item \code{nSlot}:            スロット番号
   \item \code{nNumRetrial}:      ブロック内の再割付試行のうち、調査対象者が1回以上出現した試行の数
   \item \code{gSumProp}: 　      ブロック内の再割付試行のうち、調査対象者が1回以上出現した試行を通じた、割付率の合計
   \item \code{gSumSqProp}:       ブロック内の再割付試行のうち、調査対象者が1回以上出現した試行を通じた、割付率の二乗の合計
   }
}
\description{
ウェイト算出のためのシミュレーションを実行する
}
\details{
bAPPEND==TRUEで、sDBPATHが存在するがテーブルsDBTABLE_CAT,
   sDBTABLE_SLOTが存在しない場合はエラーとなる。
}
