% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign.R
\name{sub_checkcat}
\alias{sub_checkcat}
\title{Internal: Check assigned categories}
\usage{
sub_checkcat(
  abCatNo_Use,
  abCatNo_Open,
  anCatNo_OpenCount,
  anCatNo_AllCount,
  nCAT_MAX,
  sCAT_TYPE,
  sCAT_FILTER,
  sCAT_ORDER,
  sCAT_EXCLUDE,
  bDEBUG
)
}
\arguments{
\item{abCatNo_Use}{a binary vector.
カテゴリの番号から割付可能かどうかを引くベクトル。}

\item{abCatNo_Open}{a binary vector.
カテゴリの番号からオープンかどうかを引くベクトル。}

\item{anCatNo_OpenCount}{an integer vector.
カテゴリの番号からオープンスロット数を引くベクトル。}

\item{anCatNo_AllCount}{an integer vector.
カテゴリの番号からスロット数を引くベクトル。}

\item{nCAT_MAX}{整数。
ある対象者に割り付けるカテゴリ数の上限。}

\item{sCAT_TYPE}{文字列。
カテゴリ割付タイプ。}

\item{sCAT_FILTER}{文字列。
カテゴリ割付の際の絞り込み条件。}

\item{sCAT_ORDER}{文字列。
カテゴリ割付の際の順序付け条件。}

\item{sCAT_EXCLUDE}{文字列。
カテゴリ割付の際の除外条件。}

\item{bDEBUG}{as logical.
デバッグモード}
}
\value{
a list of integer vectors.
   要素は割付カテゴリのある組み合わせ。ソート済み。
}
\description{
Internal. ある対象者について、ありうる割付カテゴリの組み合わせを
          すべて返す。
          checkSurvey()からコールされる。
}
\keyword{internal}
