% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign.R
\name{sub_checkslot}
\alias{sub_checkslot}
\title{Internal: Check assigned slots}
\usage{
sub_checkslot(
  abSlotNo_Hit,
  abSlotNo_Open,
  anSlotNo_Count,
  anSlotNo_Request,
  nSLOT_MAX,
  asSLOT_ASSIGN,
  bDEBUG
)
}
\arguments{
\item{abSlotNo_Hit}{an binary vector.
スロット番号から割付可能かどうかを引くベクトル}

\item{abSlotNo_Open}{an binary vector.
スロット番号からオープンかどうかを引くベクトル}

\item{anSlotNo_Count}{an integer vector.
スロット番号から獲得票数を引くベクトル}

\item{anSlotNo_Request}{an integer vector.
スロット番号から目標票数を引くベクトル}

\item{nSLOT_MAX}{整数。
ある対象者に割り付けるスロット数の上限。}

\item{asSLOT_ASSIGN}{文字列。
スロット割付方法。指定は必須。詳細はvignetteを参照。}

\item{bDEBUG}{as logical.
デバッグモード}
}
\value{
a list of integer vectors.
   要素は割付スロットの組み合わせ。ソート済み。
}
\description{
Internal. ある対象者のあるカテゴリについて、
          ありうる割付スロットの組み合わせをすべて返す。
          checkSurvey()からコールされる。
}
\keyword{internal}
