% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{sub_makeSubjectCat_Alt}
\alias{sub_makeSubjectCat_Alt}
\title{Internal: make a table to extract a set of alternative persons from person and category}
\usage{
sub_makeSubjectCat_Alt(mbCAT, lSLOT, sVERBOSE = c("simple", "none", "detail"))
}
\arguments{
\item{mbCAT}{整数行列。
調査対象者のカテゴリ割付可能性。
\code{mbCAT[i,j]}は以下を表す。欠損不可。
\itemize{
\item 1: 調査対象者iはカテゴリjについて割付可能。
\item 0: 調査対象者iはカテゴリjについて割付不能。
}}

\item{lSLOT}{整数行列のリスト。
調査対象者のスロット割付可能性。
\code{lSLOT[[j]][i,k]}は以下を表す。
\itemize{
\item 1: 調査対象者iはカテゴリjのスロットkについて割付可能。
\item 0: 調査対象者iはカテゴリjのスロットkについて割付不能。
}
欠損についてはDetailsを参照。}

\item{sVERBOSE}{a string.
画面表示レベル。}
}
\value{
リスト。以下の要素を持つ。
\itemize{
\item \code{manSubjectCat_AltSet}
      整数ベクトルの行列。行は対象者(mbCATの行),
      列はカテゴリ(mbCATの列), 要素はその対象者のスロット割付可能性を調べるときに
      参照すべき対象者番号のベクトル。mbCAT[i,j] == 1のときは長さ1以上のベクトル、
      mbCAT[i,j] == 0のときはNULL.
\item \code{mnSubjectCat_AltStatus}
      整数ベクトルの行列。行は対象者(mbCATの行),
      列はカテゴリ(mbCATの列), 要素は以下のいずれか:
      \itemize{
        \item NA: 割付不能カテゴリ
        \item 0: スロット割付可能性が既知
        \item 1: スロット割付可能性が未知であるためカテゴリ割付可能性が同じ対象者で代替
        \item 2: スロット割付可能性が未知であるため割付カテゴリ数が同じ対象者で代替
        \item 3: スロット割付可能性が未知であるためランダムに選んだ対象者で代替
      }
 }
}
\description{
Internal: 対象者番号とカテゴリから, 代替する対象者番号群を引く表を作る
          makeSurvey()からコールされる
}
\keyword{internal}
