% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias.R
\name{sub_makeSurvey_from_trial}
\alias{sub_makeSurvey_from_trial}
\title{Internal: make virtual survey data based on the returned value from execTrials()}
\usage{
sub_makeSurvey_from_trial(
  lPOP,
  lSETTING,
  dfSubject,
  bUSE_INFO_UNASSIGNED_CAT,
  sVERBOSE = c("simple", "detail", "none")
)
}
\arguments{
\item{lPOP}{`popdata`クラスのオブジェクト。
母集団データ。\code{\link{makePop}}で生成する。}

\item{lSETTING}{`assignsetting`クラスのオブジェクト。
割付のセッティング。\code{\link{makeSetting}} で生成する。}

\item{dfSubject}{データフレーム。
execTrials()の返し値。}

\item{sVERBOSE}{文字列。
画面表示レベル。}
}
\value{
makeSurvey()の返し値。
}
\description{
Internal: 割付シミュレーションの結果に基づいて仮想的な調査データをつくる。
          simBias()からコールされる。
}
\keyword{internal}
