% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight.R
\name{trimWeight}
\alias{trimWeight}
\title{trim Weight.}
\usage{
trimWeight(agWEIGHT, gLIMIT = 5)
}
\arguments{
\item{agWEIGHT}{数値ベクトル。ウェイト。}

\item{gLIMIT}{切り詰める限界。}
}
\value{
数値ベクトル。agWEIGHTが上限gLIMIT, 下限1/gLIMITの範囲に
収まっていたらそのまま返す。そうでないときは、
この範囲に切り詰め、かつ平均を1としたベクトルを返す。
}
\description{
ウェイトを切り詰める
}
