% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign.R
\name{execAssign}
\alias{execAssign}
\title{Internal: Assign participants into 'categories' and 'slots'}
\usage{
execAssign(
  anSEQ_PERSON,
  mnSEQ_CAT_PERSON,
  mbPERSON_CAT_USE,
  lPERSON_SLOT_HIT,
  lSETTING,
  sVERBOSE = c("simple", "detail", "none")
)
}
\arguments{
\item{anSEQ_PERSON}{an integer vector.
\code{anSEQ_PERSON[i]}は、\code{i}番目に調査に参加する対象者の
\code{mbPERSON_CAT_USE}上の行番号。欠損不可。}

\item{mnSEQ_CAT_PERSON}{an integer matrix.
\code{mnSEQ_CAT_PERSON[i,j]}は、\code{i}番目に調査に参加する
対象者の、カテゴリ\code{j}のスロット割付可能性を調べる際に参照すべき、
\code{lPERSON_SLOT_HIT[[j]]}の行番号。
欠損を許すが、
\code{mbPERSON_CAT_USE[anSEQ_PERSON[i],j] == 1}のとき、\code{mnSEQ_CAT_PERSON[i,j]}の
欠損は不可。}

\item{mbPERSON_CAT_USE}{an binary matrix.
対象者のカテゴリ割付可能性を表す。\code{mbPERSON_CAT_USE[i,j]}は以下を表す。
\itemize{
\item 1: 調査対象者iはカテゴリjについて割付可能。
\item 0: 調査対象者iはカテゴリjについて割付不能。
}
欠損を許すが、\code{mbPERSON_CAT_USE[anSEQ_PERSON, ]}における欠損は不可。}

\item{lPERSON_SLOT_HIT}{a list of binary matrices.
対象者のスロット割付可能性を表す。
要素jの行列の要素(i,k)は以下を表す。
\itemize{
\item 1: 調査対象者iはカテゴリjのスロットkについて割付可能。
\item 0: 調査対象者iはカテゴリjのスロットkについて割付不能。
}
欠損を許すが、
\code{mbPERSON_CAT_USE[i,j] == 1}のとき、
\code{lPERSON_SLOT_HIT[[j]][i, ]}における欠損は不可。}

\item{lSETTING}{`assignsetting`クラスのオブジェクト。}

\item{sVERBOSE}{a string.
画面表示レベル。}
}
\value{
an integer matrix.
   行は対象者。0行かもしれないことに注意。
   行名はmbPERSON_CAT_USEの該当行の行名。
   列は、左から順に以下の通り。ただし、
   カテゴリ番号とは\code{mbPERSON_CAT_USE}における列番号を指し、
   スロット番号とは\code{lPERSON_SLOT_HIT}の当該カテゴリにおける列番号を指す。
   \itemize{
   \item \code{SEQ}: 1             調査参加順(連番)
   \item \code{nPerson}:           対象者番号
                                   (\code{mbPERSON_CAT_USE}上の行番号。それが呼び出し元にとって
                                   なにを意味しているのかは関知しない)
   \item \code{nCat_1}:            割付カテゴリ\code{1}のカテゴリ番号、ないし\code{NA}
   \item ...
   \item \code{nCat_}(nCAT_MAX):   割付カテゴリ\code{nCAT_MAX}のカテゴリ番号、ないし\code{NA}
   \item \code{nParentCat}:        割付スロットが属するカテゴリ番号、ないし\code{NA}
   \item \code{nSlot_1}:           割付スロット\code{1}のスロット番号、ないし\code{NA}
   \item ...
   \item \code{nSlot_}(nSLOT_MAX): 割付スロット\code{nSLOT_MAX}のスロット番号、ないし\code{NA}
   }
}
\description{
Internal: 対象者をカテゴリ・スロットに割り付ける。
}
\keyword{internal}
