% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial.R
\name{execRetrials}
\alias{execRetrials}
\title{Internal: run simulations of retrials}
\usage{
execRetrials(
  lSURVEY,
  bREDRAW,
  nBLOCKSIZE,
  nNUMBLOCK,
  bPARALLEL,
  sLOGFILE,
  sVERBOSE
)
}
\arguments{
\item{lSURVEY}{a `surveydata`-class object.
調査データ。\code{\link{makeSurvey}}で生成する。}

\item{bREDRAW}{an boolean.
復元抽出するか。FALSEにすると、lSURVEY上での対象者順を固定する。}

\item{nBLOCKSIZE}{an integer.
ブロック内の再割付試行数。}

\item{nNUMBLOCK}{an integer.
実行する再割付試行のブロック数。}

\item{bPARALLEL}{a logical.
並列処理するか。}

\item{sLOGFILE}{a string.
並列処理する場合のログファイル(フルパス)。NULLだとログを出さない。}

\item{sVERBOSE}{文字列。
画面表示レベル。}
}
\value{
データフレームのリスト。要素は次の通り:
   \itemize{
   \item \code{dfStat_Cat}: カテゴリ割付頻度
   \item \code{dfStat_Slot}: スロット割付頻度
   }

   カテゴリ割付頻度: 行は調査対象者x割付カテゴリを表す。
   列は以下のとおり(順不同):
   \itemize{
   \item \code{nBlock}:           ブロック番号
   \item \code{nBlockSize}:       ブロックサイズ(ブロック内の再割付試行数)
   \item \code{nSubject}:         調査対象者番号
   \item \code{nCat}:             カテゴリ番号
   \item \code{nCount_Subject}:    ブロック内の再割付試行で調査対象者が出現した回数
   \item \code{nCount_SubjectCat}: ブロック内の再割付試行で調査対象者が出現しカテゴリに割り付けられた回数
   }

   スロット割付頻度: 行は調査対象者x割付スロットを表す。
   列は以下のとおり(順不同):
   \itemize{
   \item \code{nBlock}:           ブロック番号
   \item \code{nBlockSize}:       ブロックサイズ(ブロック内の再割付試行数)
   \item \code{nSubject}:         調査対象者番号
   \item \code{nCat}:             スロットが属するカテゴリ番号
   \item \code{nSlot}:            スロット番号
   \item \code{nCount_Subject}:   ブロック内の再割付試行で調査対象者が出現した回数
   \item \code{nCount_SubjectSlot}: ブロック内の再割付試行で調査対象者が出現しスロットに割り付けられた回数
   }
}
\description{
Internal: 再割付シミュレーションを行う.
simWeight(), simBias()からコールされる。
}
\keyword{internal}
