% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial.R
\name{makeRetrial}
\alias{makeRetrial}
\title{Internal: run execAssign() once}
\usage{
makeRetrial(lSURVEY, bREDRAW, nMAXSIZE, sVERBOSE)
}
\arguments{
\item{lSURVEY}{a `surveydata`-class object.
survey data. See the document of `makeSurvey` for details.}

\item{bREDRAW}{an boolean.
復元抽出するか。FALSEにすると、lSURVEY上での対象者順を固定する。}

\item{nMAXSIZE}{an integer.
復元抽出する対象者の最大数。bREDRAW=FALSEの場合は無視される}

\item{sVERBOSE}{文字列。画面表示レベル。}
}
\value{
a data frame.
   列は以下のとおり(順不同):
   \itemize{
   \item \code{SEQ}:              調査参加順(連番)
   \item \code{nPerson}:          対象者番号 (\code{lSURVEY$mbCAT}上の行番号)
                                  母集団メンバーの番号ではないことに注意
   \item \code{nCat_1}:           割付カテゴリ\code{1}のカテゴリ番号、ないし\code{NA}
   \item ...
   \item \code{nCat_}(nMAXCAT):   割付カテゴリ\code{lSURVEY$nMAXCAT}のカテゴリ番号、ないし\code{NA}
   \item \code{nCat}:             割付スロットが属するカテゴリ番号、ないし\code{NA}
   \item \code{nSlot_1}:          割付スロット\code{1}のスロット番号、ないし\code{NA}
   \item ...
   \item \code{nSlot_}(nMAXSLOT): 割付スロット\code{lSURVEY$nMAXSLOT}のスロット番号、ないし\code{NA}
   }
}
\description{
Internal: 再割付試行を1試行実行する。
makeRetrialBlock()からコールされる。
}
\keyword{internal}
