% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial.R
\name{makeRetrialBlock}
\alias{makeRetrialBlock}
\title{Internal: run execAssign() repeatedly to compute weight}
\usage{
makeRetrialBlock(
  lSURVEY,
  nBLOCKSIZE,
  bREDRAW,
  nMAXSIZE,
  bPARALLEL,
  sLOGFILE,
  sVERBOSE
)
}
\arguments{
\item{lSURVEY}{a `surveydata`-class object.
survey data. See the document of `makeSurvey` for details.}

\item{nBLOCKSIZE}{整数。
試行ブロック内の試行数}

\item{bREDRAW}{an boolean.
復元抽出するか。FALSEにすると、lSURVEY上での対象者順を固定する。}

\item{nMAXSIZE}{an integer.
復元抽出する対象者の最大数。bREDRAW=FALSEの場合は無視される}

\item{bPARALLEL}{a logical
並列処理するか}

\item{sLOGFILE}{a string.
並列処理する場合のログファイルの名前. NULLだとログを出さない}

\item{sVERBOSE}{文字列。画面表示レベル。}
}
\value{
a data frame.
   列は以下のとおり(順不同):
   \itemize{
   \item \code{nRetrial}:         再割付試行番号
   \item \code{nSubject}:         対象者番号 (\code{lSURVEY$mbCAT}上の行番号)
   \item \code{SEQ}:              再割付試行における調査参加順(連番)
   \item \code{nCat_1}:           割付カテゴリ\code{1}のカテゴリ番号、ないし\code{NA}
   \item ...
   \item \code{nCat_}(nMAXCAT):   割付カテゴリ\code{lSURVEY$nMAXCAT}のカテゴリ番号、ないし\code{NA}
   \item \code{nCat}:             割付スロットが属するカテゴリ番号、ないし\code{NA}
   \item \code{nSlot_1}:          割付スロット\code{1}のスロット番号、ないし\code{NA}
   \item ...
   \item \code{nSlot_}(nMAXSLOT): 割付スロット\code{lSURVEY$nMAXSLOT}のスロット番号、ないし\code{NA}
   }
}
\description{
Internal: 再割付試行を1ブロック実行する。execRetrials()からコールされる。
}
\keyword{internal}
