% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{makeSetting}
\alias{makeSetting}
\title{make setting}
\usage{
makeSetting(
  lSLOT_REQUEST,
  nCAT_MAX,
  sCAT_TYPE = c("adaptive", "nonadaptive"),
  sCAT_FILTER = c("all", "open"),
  sCAT_ORDER = c("random", "openclosed", "shortnum", "shortratio"),
  sCAT_EXCLUDE = c("none", "allclosed"),
  nSLOT_MAX,
  sSLOT_TYPE = c("adaptive", "nonadaptive"),
  sSLOT_FILTER = c("all", "open"),
  sSLOT_ORDER = c("random", "openclosed", "shortnum", "shortratio"),
  sSLOT_EXCLUDE = c("none", "allclosed"),
  nSUBJECT_MAX = 0,
  nCATSUBJECT_MAX = 0,
  nSLOTSUBJECT_MAX = 0,
  sVERBOSE = c("simple", "none", "detail")
)
}
\arguments{
\item{lSLOT_REQUEST}{a list of (named) integer vectors.
各スロットに割り付ける対象者数の目標。
要素jのベクトルの要素kは, カテゴリjのスロットkに割り付ける対象者の下限を表す。
欠損不可。
要素に名前がついている場合、スロット名とみなされるので、
名前を付けるならばすべての要素に重複なく名前をつけること。}

\item{nCAT_MAX}{a integer.
ある対象者に割り付けるカテゴリ数の上限。}

\item{sCAT_TYPE}{a string.
カテゴリ割付タイプ。詳細はvignetteを参照。}

\item{sCAT_FILTER}{a string.
カテゴリ割付の際の絞り込み条件。詳細はvignetteを参照。}

\item{sCAT_ORDER}{a string.
カテゴリ割付の際の順序付け条件。詳細はvignetteを参照。}

\item{sCAT_EXCLUDE}{a string.
カテゴリ割付の際の除外条件。詳細はvignetteを参照。}

\item{nSLOT_MAX}{an integer.
ある対象者に割り付けるスロット数の上限。}

\item{sSLOT_TYPE}{a string.
スロット割付タイプ。詳細はvignetteを参照。}

\item{sSLOT_FILTER}{a string.
スロット割付の際の絞り込み条件。詳細はvignetteを参照。}

\item{sSLOT_ORDER}{a string.
スロット割付の際の順序付け条件。詳細はvignetteを参照。}

\item{sSLOT_EXCLUDE}{a string.
スロット割付の際の除外条件。指定は必須。詳細はvignetteを参照。}

\item{nSUBJECT_MAX}{an integer.
抽出する対象者数の上限。0以上の整数。
0より大きい値が指定された場合、対象者抽出は、
対象者数がその値に達したときに中断される。}

\item{nCATSUBJECT_MAX}{an integer.
カテゴリ割付対象者数の上限。0以上の整数。
0より大きい値が指定された場合、対象者抽出は、
カテゴリ割付が生じた対象者数がその値に達したときに中断される。}

\item{nSLOTSUBJECT_MAX}{an integer.
スロット割付対象者数の上限。0以上の整数。
0より大きい値が指定された場合、対象者抽出は、
スロット割付が生じた対象者数がその値に達したときに中断される。}

\item{sVERBOSE}{a string.
画面表示レベル。}
}
\value{
an object of `assignsetting` class.
        その実体は、この関数の引数を要素として持つリスト。
        ただし、引数lSLOT_REQUESTの要素に名前がついていない場合は、
        名前"Slot_(j)_(k)"が付与される。
}
\description{
割付のセッティングを作成する。
}
\details{
対象者数が母集団サイズの10倍を超えても対象者抽出が終わらなかった場合は
   エラーとなる。
}
\examples{
lSetting <- makeSetting(
  lSLOT_REQUEST = lapply(1:10, function(x) rep(100, 10)),
  nCAT_MAX      = 1,
  sCAT_TYPE     = 'adaptive',
  sCAT_FILTER   = 'open',
  sCAT_ORDER    = 'shortnum',
  sCAT_EXCLUDE  = 'allclosed',
  nSLOT_MAX     = 2,
  sSLOT_TYPE    = 'adaptive',
  sSLOT_FILTER  = 'open',
  sSLOT_ORDER   = 'shortnum',
  sSLOT_EXCLUDE = 'allclosed',
  sVERBOSE      = "detail"
)
}
