% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{makeSetting}
\alias{makeSetting}
\title{make setting}
\usage{
makeSetting(
  lSLOT_REQUEST,
  nCAT_MAX,
  nSLOT_MAX,
  sCAT_TYPE = c("adaptive", "nonadaptive"),
  sCAT_FILTER = c("all", "open"),
  sCAT_ORDER = c("random", "openclosed", "shortnum", "shortratio"),
  sCAT_EXCLUDE = c("none", "allclosed"),
  sSLOT_TYPE = c("adaptive", "nonadaptive"),
  sSLOT_FILTER = c("all", "open"),
  sSLOT_ORDER = c("random", "openclosed", "shortnum", "shortratio"),
  sSLOT_EXCLUDE = c("none", "allclosed"),
  nSUBJECT_MAX = 0,
  sVERBOSE = c("simple", "none", "detail")
)
}
\arguments{
\item{lSLOT_REQUEST}{a list of (named) integer vectors.
各スロットに割り付ける対象者数の目標。
要素jのベクトルの要素kは, カテゴリjのスロットkに割り付ける対象者の下限を表す。
欠損不可。
要素に名前がついている場合、スロット名とみなされるので、
名前を付けるならばすべての要素に重複なく名前をつけること。}

\item{nCAT_MAX}{a integer.
ある対象者に割り付けるカテゴリ数の上限。}

\item{nSLOT_MAX}{an integer.
ある対象者に割り付けるスロット数の上限。}

\item{sCAT_TYPE}{a string.
カテゴリ割付タイプ。詳細はvignetteを参照。}

\item{sCAT_FILTER}{a string.
カテゴリ割付の際の絞り込み条件。詳細はvignetteを参照。}

\item{sCAT_ORDER}{a string.
カテゴリ割付の際の順序付け条件。詳細はvignetteを参照。}

\item{sCAT_EXCLUDE}{a string.
カテゴリ割付の際の除外条件。詳細はvignetteを参照。}

\item{sSLOT_TYPE}{a string.
スロット割付タイプ。詳細はvignetteを参照。}

\item{sSLOT_FILTER}{a string.
スロット割付の際の絞り込み条件。詳細はvignetteを参照。}

\item{sSLOT_ORDER}{a string.
スロット割付の際の順序付け条件。詳細はvignetteを参照。}

\item{sSLOT_EXCLUDE}{a string.
スロット割付の際の除外条件。指定は必須。詳細はvignetteを参照。}

\item{nSUBJECT_MAX}{an integer.
   抽出する対象者数の上限。0以上の整数。

   0より大きい値が指定された場合、
   対象者抽出は、
   対象者数がnSUBJECT_MAXに達するか、すべてのスロットの割付対象者数が
   lSLOT_REQUESTに設定された目標に達したときに終了する。

   0が指定された場合、
   対象者抽出は、すべてのスロットの割付対象者数が目標に達したときにのみ正常終了する。
   対象者数が母集団サイズの10倍を超えても
   割付対象者数が目標に到達しなかったときはエラーとなる。}

\item{sVERBOSE}{a string.
画面表示レベル。}
}
\value{
an object of `assignsetting` class.
        その実体は、この関数の引数を要素として持つリスト。
        ただし、引数lSLOT_REQUESTの要素に名前がついていない場合は、
        名前"Slot_(j)_(k)"が付与される。
}
\description{
割付のセッティングを作成する。
}
