% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial.R
\name{makeStat_Cat}
\alias{makeStat_Cat}
\title{Interal: make statistics for categories}
\usage{
makeStat_Cat(dfReplicateBlock, lSURVEY, bVERBOSE)
}
\arguments{
\item{dfReplicateBlock}{a data frame
対象者のカテゴリ・ウェイトへの割付を反復実行した結果。
以下の列を持つと期待される
 \itemize{
 \item \code{nSubject}:          対象者番号
 \item \code{nCat_1}:           割付カテゴリ\code{1}のカテゴリ番号、ないし\code{NA}
 \item ...
 \item \code{nCat_}(nMAXCAT):   割付カテゴリ\code{nMAXCAT}のカテゴリ番号、ないし\code{NA}
 \item \code{nCat}:             割付スロットが属するカテゴリ番号、ないし\code{NA}
 \item \code{nSlot_1}:          割付スロット\code{1}のスロット番号、ないし\code{NA}
 \item ...
 \item \code{nSlot_}(nMAXSLOT): 割付スロット\code{nMAXSLOT}のスロット番号、ないし\code{NA}
 }}

\item{lSURVEY}{a `surveydata`-class object.
survey data. See the document of `makeSurvey` for details.}

\item{bVERBOSE}{a logical.
詳細を画面表示する}
}
\value{
a data frame
   行は対象者x割付カテゴリ
   \itemize{
   \item \code{nSubject}          対象者番号
   \item \code{nCat}             親カテゴリ番号
   \item \code{nCount_SubjectCat} その人にそのカテゴリが割り付けられた回数
   \item \code{nCount_Subject}    その人の出現回数
   }
}
\description{
Internal: カテゴリ割付頻度を集計する。execRetrials()からコールされる。
}
\keyword{internal}
