% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{makeSurvey}
\alias{makeSurvey}
\title{make survey data.}
\usage{
makeSurvey(
  mbCAT,
  lSLOT,
  lSLOT_REQUEST,
  mnASSIGNCAT,
  anPARENTCAT,
  mnASSIGNSLOT,
  sCAT_TYPE = c("adaptive", "nonadaptive"),
  sCAT_FILTER = c("all", "open"),
  sCAT_ORDER = c("random", "openclosed", "shortnum", "shortratio"),
  sCAT_EXCLUDE = c("none", "allclosed"),
  sSLOT_TYPE = c("adaptive", "nonadaptive"),
  sSLOT_FILTER = c("all", "open"),
  sSLOT_ORDER = c("random", "openclosed", "shortnum", "shortratio"),
  sSLOT_EXCLUDE = c("none", "allclosed"),
  nSUBJECT_MAX = NULL,
  sVERBOSE = c("simple", "none", "detail")
)
}
\arguments{
\item{mbCAT}{整数行列。
調査対象者のカテゴリ割付可能性。
\code{mbCAT[i,j]}は以下を表す。欠損不可。
\itemize{
\item 1: 調査対象者iはカテゴリjについて割付可能。
\item 0: 調査対象者iはカテゴリjについて割付不能。
}
列に名前を付けるとカテゴリ名とみなされる。
名前を付けるならばすべての列に重複なく名前をつけること。}

\item{lSLOT}{整数行列のリスト。
調査対象者のスロット割付可能性。
\code{lSLOT[[j]][i,k]}は以下を表す。
\itemize{
\item 1: 調査対象者iはカテゴリjのスロットkについて割付可能。
\item 0: 調査対象者iはカテゴリjのスロットkについて割付不能。
}
欠損についてはDetailsを参照。
列に名前を付けるとスロット名とみなされる。
名前を付けるならばすべての列に重複なく名前をつけること。}

\item{lSLOT_REQUEST}{整数ベクトルのリスト。
各スロットに割り付ける対象者数の下限。
要素jのベクトルの要素kは, カテゴリjのスロットkに割り付ける対象者の下限を表す。
欠損なし。
要素の列名はlSLOTの要素の列名と一致していること。}

\item{mnASSIGNCAT}{整数行列。
調査対象者の割付カテゴリ。列数は、割付カテゴリ数の上限。
\code{mnASSIGNCAT[i,j]}は、調査対象者iのj番目の割付カテゴリ番号(mbCATの列番号)を表す。
調査対象者iの割付カテゴリ数が列数より小さい場合はNAを埋める。
列順には意味がない。}

\item{anPARENTCAT}{整数ベクトル。
調査対象者の割付スロットが属するカテゴリ。
\code{anPARENTCAT[i]}は、調査対象者iの割付スロットが属するカテゴリ番号(mbCATの列番号)を表す。
調査対象者iがスロットに割り付けられなかった場合はNAを埋める。}

\item{mnASSIGNSLOT}{整数行列。
調査対象者の割付スロット。列数は、割付スロット数の上限。
\code{mnASSIGNSLOT[i,j]}は、調査対象者iの割付スロット番号(lSLOT[[anPARENTCAT[i]]]の列番号)を
表す。調査対象者iの割付スロット数が列数より小さい場合はNAを埋める。
列順には意味がない。}

\item{sCAT_TYPE}{文字列。
カテゴリ割付タイプ。指定は必須。詳細はvignetteを参照。}

\item{sCAT_FILTER}{文字列。
カテゴリ割付の際の絞り込み条件。指定は必須。詳細はvignetteを参照。}

\item{sCAT_ORDER}{文字列。
カテゴリ割付の際の順序付け条件。指定は必須。詳細はvignetteを参照。}

\item{sCAT_EXCLUDE}{文字列。
カテゴリ割付の際の除外条件。指定は必須。詳細はvignetteを参照。}

\item{sSLOT_TYPE}{文字列。
スロット割付タイプ。指定は必須。詳細はvignetteを参照。}

\item{sSLOT_FILTER}{文字列。
スロット割付の際の絞り込み条件。指定は必須。詳細はvignetteを参照。}

\item{sSLOT_ORDER}{文字列。
スロット割付の際の順序付け条件。指定は必須。詳細はvignetteを参照。}

\item{sSLOT_EXCLUDE}{文字列。
スロット割付の際の除外条件。指定は必須。詳細はvignetteを参照。}

\item{nSUBJECT_MAX}{整数。
   抽出する対象者数の上限。

  \code{nSUBJECT_MAX}が指定された場合、
   対象者抽出は、
   対象者数がnSUBJECT_MAXに達するか、すべてのスロットの割付対象者数が
   lSLOT_REQUESTに設定された目標に達したときに終了する。

  \code{nSUBJECT_MAX}がNULLの場合、
   対象者抽出は、すべてのスロットの割付対象者数が目標に達したときにのみ正常終了する。
   対象者数が母集団サイズの10倍を超えても
   割付対象者数が目標に到達しなかったときはエラーとなる。}

\item{sVERBOSE}{文字列。
画面表示レベル。}
}
\value{
`surveydata`クラスのオブジェクト。その実体は以下の要素を持つリスト。
\itemize{
\item \code{mbCAT}         整数行列。引数mbCATの値。
                           列名がついていなかった場合は、
                           列名"Cat_(j)"が付与される。
\item \code{lSLOT}         整数行列のリスト。引数lSLOTの値。ただし、割付不能カテゴリの
                           スロット割付可能性はすべてNAに変更される。また、要素の行にひとつでもNAがあったら
                           その行はみなNAに変更される。
                           列名がついていなかった場合は、
                           列名"Slot_(j)_(k)"が付与される。
\item \code{lSLOT_REQUEST} 整数ベクトルのリスト。引数lSLOT_REQUESTの値。
                           名前がついていなかった場合は、
                           名前"Slot_(j)_(k)"が付与される。
\item \code{sCAT_TYPE}     文字列。引数sCAT_TYPEの値。
\item \code{sCAT_FILTER}   文字列。引数sCAT_FILTERの値。
\item \code{sCAT_ORDER}    文字列。引数sCAT_ORDERの値。
\item \code{sCAT_EXCLUDE}  文字列。引数sCAT_EXCLUDEの値。
\item \code{sSLOT_TYPE}    文字列。引数sSLOT_TYPEの値。
\item \code{sSLOT_FILTER}  文字列。引数sSLOT_FILTERの値。
\item \code{sSLOT_ORDER}   文字列。引数sSLOT_ORDERの値。
\item \code{sSLOT_EXCLUDE} 文字列。引数sSLOT_EXCLUDEの値。
\item \code{mnASSIGNCAT}   整数行列。引数mnASSIGNCATの値。
\item \code{anPARENTCAT}   整数ベクトル。引数anPARENTCATの値。
\item \code{mnASSIGNSLOT}  整数行列。引数mnASSIGNSLOTの値。
\item \code{nSUBJECT_MAX}  整数。引数nSUBJECT_MAXの値。
\item \code{manSubjectCat_AltSubject} 整数ベクトルを要素とする行列。行は対象者(mbCATの行),
                                      列はカテゴリ(mbCATの列),
                                      要素はその対象者のスロット割付可能性を調べるときに
                                      参照すべき対象者番号のベクトル。
\item \code{mnSubjectCat_AltStatus}   整数ベクトルの行列。行は対象者(mbCATの行),
                                      列はカテゴリ(mbCATの列), 要素は以下のいずれか:
\itemize{
\item NA: 割付不能カテゴリ
\item 0: スロット割付可能性が既知
\item 1: スロット割付可能性が未知であるためカテゴリ割付可能性が同じ対象者で代替
\item 2: スロット割付可能性が未知であるため割付カテゴリ数が同じ対象者で代替
\item 3: スロット割付可能性が未知であるためランダムに選んだ対象者で代替
}
}
}
\description{
調査データを作成する.
}
\details{
\code{lSLOT}の欠損は以下のように扱われる。
  \itemize{
  \item Case 1. \code{mbCAT[i,j] == 0}のとき、
  \code{lSLOT[[j]][i, ]}は無視される。すなわち、
  ある対象者があるカテゴリに対して割付不能な時、
  その対象者のそのカテゴリに属するスロットへの割付可能性は未知として扱われ、
  指定しても無視される。
  \item Case 2. \code{mbCAT[i,j] == 1}であり\code{j} が\code{mnASSIGNCAT[i,]}に含まれているとき、
  \code{lSLOT[[j]][i, ]}の欠損は許されない。すなわち、
  ある対象者があるカテゴリに割り付けられた場合、
  その対象者のそのカテゴリに属するスロットへの割付可能性は既知であるはずなので、
  指定しなければならない。
  \item Case 3. \code{mbCAT[i,j] == 1}であり\code{j} が\code{mnASSIGNCAT[i,]}に含まれていないとき、
  \code{lSLOT[[j]][i, ]}の欠損はある程度まで許容される。すなわち、
  ある対象者があるカテゴリに割付可能であったが割り付けられなかった場合、
  その対象者のそのカテゴリに属するスロットへの割付可能性は、分かっている限りにおいて
  指定すればよい。ある行に欠損がある場合、その行はすべて欠損とみなされる。
  }

  Case 3.に相当する調査対象者も、
  再割付シミュレーションにおいてカテゴリ j に割り付けられる可能性がある。
  そのため、jに属するスロットへの割付可能性が欠損である調査対象者については、
  それらの欠損を補完する必要が生じる。このとき、
  \code{SurveyAssign}パッケージによるシミュレーションでは、
  調査対象者 i の割付可能性を
  他の調査対象者 i' の割付可能性によって置き換える。
  i' はシミュレーション試行ごとに、次の手順で選ばれる。
  \itemize{
  \item Step 1. jに属するスロットへの割付可能性が既知であり、
  かつJ個のカテゴリへの割付可能性が i と同一である人の中からランダムに選ぶ。
  \item Step 2. みつからなかった場合は、jに属するスロットへの割付可能性が既知であり、
  かつ割付可能なカテゴリの数が同一である人の中からランダムに選ぶ。
  \item Step 3. みつからなかった場合は、jに属するスロットへの割付可能性が既知である
  人の中からランダムに選ぶ。
  }

  \code{makeSurvey}は、今後のシミュレーションの際に
  上記の手順での補完が可能かどうかをチェックし、Step 1で補完できない人が
  出現した場合には警告を表示する。
}
