% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias.R
\name{simBias}
\alias{simBias}
\title{run simulations to evaluate bias}
\usage{
simBias(
  lPOP,
  lSETTING,
  nNUMTRIAL = 1,
  nNUMRETRIAL = 1000,
  bUSE_INFO_UNASSIGNED_CAT = TRUE,
  bPARALLEL = FALSE,
  sLOGFILE = NULL,
  sDBPATH = NULL,
  sDBTABLE_SUBJECT = "subject",
  sDBTABLE_CAT = "weight_cat",
  sDBTABLE_SLOT = "weight_slot",
  bAPPEND = TRUE,
  sVERBOSE = c("simple", "detail", "none")
)
}
\arguments{
\item{lPOP}{`popdata`クラスのオブジェクト。
母集団データ。\code{\link{makePop}}で生成する。}

\item{lSETTING}{`assignsetting'クラスのオブジェクト。
割付のセッティング。\code{\link{makeSetting}}で生成する。}

\item{nNUMTRIAL}{整数。
割付試行数。}

\item{nNUMRETRIAL}{整数。
各割付試行における再割付試行数。}

\item{bUSE_INFO_UNASSIGNED_CAT}{論理値。
再割付試行において、非割付カテゴリに属するスロットへの割付可能性を利用するか。
TRUEの場合、すべての割付可能カテゴリに属するスロット割付可能性を用いて
再割付試行を行う。
FALSEの場合、割付カテゴリに属するスロット割付可能性のみを用いて
再割付試行を行う。}

\item{bPARALLEL}{論理値。
並列処理するか。}

\item{sLOGFILE}{文字列。
並列処理する場合のログファイル(フルパス)。NULLだとログを出さない。}

\item{sDBPATH}{文字列。
シミュレーションの結果を保存するSQLite DBのフルパス。}

\item{sDBTABLE_SUBJECT}{文字列。
割付シミュレーションの結果を保存するSQLite DBのテーブル名。}

\item{sDBTABLE_CAT}{文字列。
再割付シミュレーションの結果得られたカテゴリ割付頻度を保存するSQLite DBのテーブル名。}

\item{sDBTABLE_SLOT}{文字列。
再割付シミュレーションの結果得られたスロット割付頻度を保存するSQLite DBのテーブル名。}

\item{bAPPEND}{論理値。
TRUE: 結果をテーブルに追加する。FALSE: 上書きする。}

\item{sVERBOSE}{文字列。
画面表示レベル。}
}
\value{
\itemize{
     \item sDBPATHを指定しない場合は、シミュレーションの結果をデータフレームのリストとして返す。
     \item sDBPATHを指定した場合はNULLを返す。シミュレーションの結果は、
           SQLiteデータベースにテーブルとして保存される。sDBPATH が存在する場合は、
           bAPPEND==FALSEであればSDBPATHに上書きし、bAPPEND==TRUEであれば追加する。
           sDBPATHが存在しない場合は新規作成する。
   }

   シミュレーションの結果は次の3つのデータフレームからなる。
   \itemize{
     \item \code{dfSubject}:
           行はある試行のある調査参加者を表す。
           列は\code{execTrial()}の出力に試行番号 \code{nTrial}を追加したもの。
     \item \code{dfStat_Cat}: 行はある試行のあるカテゴリを表す。
           列は\code{execRetrial()}の出力のうち\code{dfStat_Cat}に試行番号 \code{nTrial}を追加したもの。
     \item \code{dfStat_Slot}: 行はある試行のあるカテゴリのあるスロットを表す。
           列は\code{execRetrial()}の出力のうち\code{dfStat_Slot}に試行番号 \code{nTrial}を追加したもの。
   }
}
\description{
バイアス評価のためのシミュレーションを実行する
}
\details{
bAPPEND==TRUEで、sDBPATHが存在し、かつ
   テーブルsDBTABLE_SUBJECT, sDBTABLE_CAT, sDBTABLE_SLOTが存在しない場合はエラーとなる。
}
