% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight.R
\name{simWeight}
\alias{simWeight}
\title{run simulations for computing weights}
\usage{
simWeight(
  lSURVEY,
  bREDRAW = TRUE,
  nBLOCKSIZE = 100,
  nNUMBLOCK = 1,
  bPARALLEL = FALSE,
  sLOGFILE = NULL,
  sDBPATH = NULL,
  sDBTABLE_CAT = "weight_cat",
  sDBTABLE_SLOT = "weight_slot",
  bAPPEND = FALSE,
  sVERBOSE = c("simple", "detail", "none")
)
}
\arguments{
\item{lSURVEY}{`surveydata`クラスのオブジェクト。
調査データ。\code{\link{makeSurvey}}で生成する。}

\item{bREDRAW}{論理値。
FALSEにすると、再割付試行においてlSURVEYの対象者順を固定する。}

\item{nBLOCKSIZE}{整数。
ブロック内の再割付試行数。}

\item{nNUMBLOCK}{整数。
実行する再割付試行のブロック数。}

\item{bPARALLEL}{論理値。
並列処理するか。}

\item{sLOGFILE}{文字列。
並列処理する場合のログファイル(フルパス)。NULLだとログを出さない。}

\item{sDBPATH}{文字列。
シミュレーションの結果を保存するSQLite DBのフルパス。
存在しない場合は作成する。}

\item{sDBTABLE_CAT}{文字列。
再割付シミュレーションの結果得られたカテゴリ割付頻度を保存するSQLite DBのテーブル名。}

\item{sDBTABLE_SLOT}{文字列。
再割付シミュレーションの結果得られたスロット割付頻度を保存するSQLite DBのテーブル名。}

\item{bAPPEND}{論理値。
結果をテーブルに{TRUE: 追加する, FALSE: 上書きする}。}

\item{sVERBOSE}{文字列。
画面表示レベル。}
}
\value{
sDBPATHを指定しない場合は、シミュレーションの結果を
   データフレームのリストとして返す。要素は次の通り:
   \itemize{
   \item \code{dfStat_Cat}: カテゴリ割付頻度
   \item \code{dfStat_Slot}: スロット割付頻度
   }

   sDBPATHを指定した場合はNULLを返す。
   シミュレーションの結果は、SQLiteデータベースにテーブルsDBTABLE_CAT,
   sDBTABLE_SLOTとして保存される。

   bAPPEND==FALSEの場合、SQLiteデータベースファイル sDBPATH
   が存在する場合は上書きし、存在しない場合は新規作成する。

   bAPPEND==TRUEの場合、SQLiteデータベースファイル sDBPATH に
   テーブルsDBTABLE_CAT, sDBTABLE_SLOT が存在する場合には
   それらのテーブルに追加し、
   ファイルが存在しない場合はファイルとテーブルを新規作成する。

   シミュレーションの結果は以下の形式を持つ。

   カテゴリ割付頻度: 行は調査対象者x割付カテゴリを表す。
   列は以下のとおり(順不同):
   \itemize{
   \item \code{nBlock}:           ブロック番号
   \item \code{nBlockSize}:       ブロックサイズ(ブロック内の再割付試行数)
   \item \code{nSubject}:         調査対象者番号
   \item \code{nCat}:             カテゴリ番号
   \item \code{nCount_Subject}:    ブロック内の再割付試行で調査対象者が出現した回数
   \item \code{nCount_SubjectCat}: ブロック内の再割付試行で調査対象者が出現しカテゴリに割り付けられた回数
   }

   スロット割付頻度: 行は調査対象者x割付スロットを表す。
   列は以下のとおり(順不同):
   \itemize{
   \item \code{nBlock}:           ブロック番号
   \item \code{nBlockSize}:       ブロックサイズ(ブロック内の再割付試行数)
   \item \code{nSubject}:         調査対象者番号
   \item \code{nCat}:             スロットが属するカテゴリ番号
   \item \code{nSlot}:            スロット番号
   \item \code{nCount_Subject}:   ブロック内の再割付試行で調査対象者が出現した回数
   \item \code{nCount_SubjectSlot}: ブロック内の再割付試行で調査対象者が出現しスロットに割り付けられた回数
   }
}
\description{
ウェイト算出のためのシミュレーションを実行する
}
\details{
bAPPEND==TRUEで、sDBPATHが存在するがテーブルsDBTABLE_CAT,
   sDBTABLE_SLOTが存在しない場合はエラーとなる。
}
