% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign.R
\name{sub_getslot}
\alias{sub_getslot}
\title{Internal: Assign a participant into slots}
\usage{
sub_getslot(
  abSlotNo_Hit,
  abSlotNo_Open,
  anSlotNo_Count,
  anSlotNo_Request,
  nSLOT_MAX,
  sSLOT_TYPE,
  sSLOT_FILTER,
  sSLOT_ORDER,
  sSLOT_EXCLUDE,
  bDEBUG
)
}
\arguments{
\item{abSlotNo_Hit}{an binary vector.
スロット番号から割付可能かどうかを引くベクトル}

\item{abSlotNo_Open}{an binary vector.
スロット番号からオープンかどうかを引くベクトル}

\item{anSlotNo_Count}{an integer vector.
スロット番号から獲得票数を引くベクトル}

\item{anSlotNo_Request}{an integer vector.
スロット番号から目標票数を引くベクトル}

\item{nSLOT_MAX}{整数。
ある対象者に割り付けるスロット数の上限。}

\item{sSLOT_TYPE}{文字列。
スロット割付タイプ。指定は必須。詳細はvignetteを参照。}

\item{sSLOT_FILTER}{文字列。
スロット割付の際の絞り込み条件。指定は必須。詳細はvignetteを参照。}

\item{sSLOT_ORDER}{文字列。
スロット割付の際の順序付け条件。指定は必須。詳細はvignetteを参照。}

\item{sSLOT_EXCLUDE}{文字列。
スロット割付の際の除外条件。指定は必須。詳細はvignetteを参照。}

\item{bDEBUG}{as logical.
デバッグモード}
}
\value{
an integer vector.
   割付スロットの番号。長さはnSLOT_MAX, 足りない分はNAを詰める
}
\description{
Internal. ある対象者について、あるカテゴリの割付スロットを決定する。
          execAssign()からコールされる。
}
\keyword{internal}
